requirements/common.txt
celery
Django==3.1.14
django-appconf
django-assets
django-celery-beat
django-celery-results
django-extensions
django-filter
django-import-export
django-jsonify
django-rest-swagger
django-structlog
django-timezone-field
djangorestframework>=3.6
drfapikey
flower>=0.9.5
markdown
natsort
python-debian
PyYAML
six>=1.9
webassets
